objId is size_t and compulsory
git-svn-id: file:///home/roberto/Git/svn@7148 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: 2d155be7a18aad023bc592f124043743994b2d2f

+++ trunk/xmipp/libraries/reconstruction/make_spectra.h
@@ -68,7 +68,7 @@ public:
 
     /** Process image.
      */
-    void processImage(const FileName &fnImg, const FileName &fnImgOut, long int objId);
+    void processImage(const FileName &fnImg, const FileName &fnImgOut, size_t objId);
 
     /** Finish processing.
      */
